backend/requirements.txt
fastapi==0.100.1
llama_cpp_python==0.1.73
passlib==1.7.4
psycopg2==2.9.6
psycopg2_binary==2.9.6
pydantic==1.10.9
python_jose==3.3.0
starlette==0.31.0
